E2E/Playwright: Reuse merged POMs in Cypress batch 2
