+++ e2e-tests/playwright/lib/src/ui/components/channels/direct_channels_modal.ts
@@ -12,6 +12,7 @@ export default class DirectChannelsModal {
     readonly results;
     readonly searchInput;
     readonly memberLimitHelpText;
+    readonly newConversationWarning;
 
     constructor(container: Locator) {
         this.container = container;
@@ -20,6 +21,9 @@ export default class DirectChannelsModal {
         this.results = container.getByTestId('more-modal-list');
         this.searchInput = container.getByRole('combobox', {name: 'Search for people'});
         this.memberLimitHelpText = container.locator('#multiSelectHelpMemberInfo');
+        this.newConversationWarning = container.getByText(
+            "This will start a new conversation. If you're adding a lot of people, consider creating a private channel instead.",
+        );
     }
 
     async toBeVisible() {
@@ -67,6 +71,17 @@ export default class DirectChannelsModal {
 
     async fillSearchInput(text: string) {
         await this.searchInput.fill(text);
+    }
+
+    getUserResult(username: string) {
+        return this.results.getByText(`@${username}`, {exact: false});
+    }
+
+    async openSelfDirectMessage(username: string) {
+        await this.fillSearchInput(username);
+        await expect(this.getUserResult(username)).toBeVisible();
+        await this.getUserResult(username).click();
+        await expect(this.container).not.toBeAttached();
     }
 
     async toHaveUserAsNthResult(user: UserProfile, index: number) {
