+++ e2e-tests/playwright/specs/functional/channels/messaging/group_message.spec.ts
@@ -4,53 +4,6 @@
 import {expect, test} from '@mattermost/playwright-lib';
 
 /**
- * @objective Verify a reaction can be added to a group message post and its action visibility adapts between desktop and mobile layouts.
- */
-test('MM-T471 add a reaction to a message in a GM', {tag: '@messaging'}, async ({pw}) => {
-    const {adminClient, team, user} = await pw.initSetup();
-    const members = await adminClient.createUsers(team.id, 2, 'gm-reaction');
-    const groupChannel = await adminClient.createGroupChannel([user.id, ...members.map((member) => member.id)]);
-
-    // # Open the group message and post a message
-    const {channelsPage, page} = await pw.testBrowser.login(user);
-    await channelsPage.gotoMessage(team.name, groupChannel.name);
-    await channelsPage.toBeVisible();
-    await channelsPage.postMessage('This is a post');
-    const post = await channelsPage.getLastPost();
-
-    // # Open the post reaction picker and choose slightly frowning face
-    await post.openReactionPicker();
-    await channelsPage.reactionEmojiPicker.clickEmoji('slightly frowning face');
-
-    // * Verify the reaction is visible with a count of one
-    const reaction = post.container.getByRole('button', {name: /slightly_frowning_face/i});
-    await expect(reaction).toBeVisible();
-    await expect(reaction).toContainText('1');
-    const addReactionButton = post.container.getByRole('button', {name: 'Add a reaction', exact: true});
-
-    // # Click the channel intro, then focus the message input to clear post focus
-    await channelsPage.centerView.channelIntro.click();
-    await channelsPage.centerView.postCreate.input.click();
-
-    // * Verify the Add Reaction action is hidden when the desktop post is not hovered
-    await expect(addReactionButton).not.toBeVisible();
-
-    // # Hover the post, then focus the message input again
-    await post.hover();
-    await expect(addReactionButton).toBeVisible();
-    await channelsPage.centerView.postCreate.input.click();
-
-    // * Verify the Add Reaction action is hidden again
-    await expect(addReactionButton).not.toBeVisible();
-
-    // # Resize the window to a mobile viewport
-    await page.setViewportSize({width: 375, height: 667});
-
-    // * Verify the Add Reaction action is visible in the mobile layout
-    await expect(addReactionButton).toBeVisible();
-});
-
-/**
  * @objective Verify a group message channel header can be added, posts a system message, and updates read state for participants.
  */
 test('MM-T472 Add a channel header to a GM', {tag: '@messaging'}, async ({pw}) => {
@@ -63,27 +16,15 @@ test('MM-T472 Add a channel header to a GM', {tag: '@messaging'}, async ({pw}) =
     const {channelsPage, page} = await pw.testBrowser.login(user);
     await channelsPage.gotoMessage(team.name, groupChannel.name);
     await channelsPage.toBeVisible();
-    const addHeaderButton = channelsPage.centerView.header.container.getByRole('button', {
-        name: 'Add a channel header',
-    });
-
     // * Verify no channel header is set
-    await expect(addHeaderButton).not.toBeVisible();
+    await expect(channelsPage.centerView.header.addChannelHeaderButton).not.toBeVisible();
 
     // # Hover the header, open the header editor, and save a header
-    await channelsPage.centerView.header.container.hover();
-    await addHeaderButton.click();
-    const modalHeading = page.getByRole('heading', {name: /Edit Header for/});
-    await expect(modalHeading).toBeVisible();
-    const headerInput = page.getByRole('textbox', {
-        name: 'Edit the text appearing next to the channel name in the header.',
-    });
-    await headerInput.fill(header);
-    await headerInput.press('Enter');
-    await expect(modalHeading).not.toBeVisible();
+    await channelsPage.centerView.header.openAddChannelHeader();
+    await channelsPage.editChannelHeaderModal.setHeaderWithEnter(header);
 
     // * Verify the header appears and a deletable system message records the update
-    await expect(channelsPage.centerView.header.container.getByText(header, {exact: true})).toBeVisible();
+    await expect(channelsPage.centerView.header.getHeaderText(header)).toBeVisible();
     const systemPost = await channelsPage.getLastPost();
     await systemPost.toContainText('updated the channel header');
     await systemPost.hover();
@@ -114,30 +55,20 @@ test('MM-T473_1 Edit GM channel header', {tag: '@messaging'}, async ({pw}) => {
     const header = 'In pursuit of peace and progress';
 
     // # Open the group message with an existing header
-    const {channelsPage, page} = await pw.testBrowser.login(user);
+    const {channelsPage} = await pw.testBrowser.login(user);
     await channelsPage.gotoMessage(team.name, groupChannel.name);
     await channelsPage.toBeVisible();
 
     // * Verify the add-header action is absent because a header is already set
-    await expect(
-        channelsPage.centerView.header.container.getByRole('button', {name: 'Add a channel header'}),
-    ).toHaveCount(0);
+    await expect(channelsPage.centerView.header.addChannelHeaderButton).toHaveCount(0);
 
     // # Open the channel menu, choose Edit Header, and save the replacement
-    await channelsPage.centerView.header.openChannelMenu();
-    await page.getByRole('menuitem', {name: 'Settings'}).hover();
-    await page.getByRole('menuitem', {name: 'Edit Header'}).click();
-    const modalHeading = page.getByRole('heading', {name: /Edit Header for/});
-    await expect(modalHeading).toBeVisible();
-    const headerInput = page.getByRole('textbox', {
-        name: 'Edit the text appearing next to the channel name in the header.',
-    });
-    await headerInput.fill(header);
-    await headerInput.press('Enter');
-    await expect(modalHeading).not.toBeVisible();
+    const channelMenu = await channelsPage.openChannelMenu();
+    await channelMenu.openEditHeader();
+    await channelsPage.editChannelHeaderModal.setHeaderWithEnter(header);
 
     // * Verify the new header appears and a system message records the update
-    await expect(channelsPage.centerView.header.container.getByText(header, {exact: true})).toBeVisible();
+    await expect(channelsPage.centerView.header.getHeaderText(header)).toBeVisible();
     await (await channelsPage.getLastPost()).toContainText('updated the channel header');
 
     // # Sign in as another group member and view Town Square
@@ -150,7 +81,7 @@ test('MM-T473_1 Edit GM channel header', {tag: '@messaging'}, async ({pw}) => {
 });
 
 /**
- * @objective Verify a username mention in a group message channel header renders as a link without highlighting the current user.
+ * @objective Verify a username mention in a group message channel header is interactive without highlighting the current user.
  */
 test('MM-T473_2 Edit GM channel header with a mention', {tag: '@messaging'}, async ({pw}) => {
     const {adminClient, team, user} = await pw.initSetup();
@@ -160,21 +91,15 @@ test('MM-T473_2 Edit GM channel header with a mention', {tag: '@messaging'}, asy
     const header = `In pursuit of peace and progress by @${user.username}`;
 
     // # Open the group message and edit the existing header
-    const {channelsPage, page} = await pw.testBrowser.login(user);
+    const {channelsPage} = await pw.testBrowser.login(user);
     await channelsPage.gotoMessage(team.name, groupChannel.name);
     await channelsPage.toBeVisible();
-    await channelsPage.centerView.header.openChannelMenu();
-    await page.getByRole('menuitem', {name: 'Settings'}).hover();
-    await page.getByRole('menuitem', {name: 'Edit Header'}).click();
-    const headerInput = page.getByRole('textbox', {
-        name: 'Edit the text appearing next to the channel name in the header.',
-    });
-    await headerInput.fill(header);
-    await page.getByRole('button', {name: 'Save', exact: true}).click();
-    await expect(page.getByRole('heading', {name: /Edit Header for/})).not.toBeVisible();
+    const channelMenu = await channelsPage.openChannelMenu();
+    await channelMenu.openEditHeader();
+    await channelsPage.editChannelHeaderModal.setHeader(header);
 
-    // * Verify the mention is a link in the header and is not highlighted for the mentioned user
-    const mention = channelsPage.centerView.header.container.locator('.mention-link');
+    // * Verify the mention opens the user's profile and is not highlighted for the mentioned user
+    const mention = channelsPage.centerView.header.getHeaderMention(`@${user.username}`);
     await expect(mention).toBeVisible();
     await expect(mention).toHaveText(`@${user.username}`);
     await expect(mention).not.toHaveClass(/mention--highlight/);
